requirements.txt
# Core runtime dependencies

# API clients
openai>=1.30.0
boto3>=1.34.0
google-genai>=1.0.0

# Config
python-dotenv>=1.0.0
